Fixing issue with std::max in 64 bit build. Since size_t changes to 64, std::max template definition fails to compare a 64 bit number and a 32 bit

+++ source/Library.Shared/Vector.inl
@@ -74,7 +74,7 @@ namespace AnonymousEngine
 
 	template <typename T>
 	Vector<T>::Vector(const std::initializer_list<T>& items) :
-		Vector(std::max(items.size(), DefaultCapacity))
+		Vector(std::max(static_cast<std::uint32_t>(items.size()), DefaultCapacity))
 	{
 		for (const auto& item : items)
 		{
